+++ ft_putstr.c
@@ -6,13 +6,13 @@
 /*   By: bgazur <bgazur@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/05 14:24:42 by bgazur            #+#    #+#             */
-/*   Updated: 2025/05/05 15:13:07 by bgazur           ###   ########.fr       */
+/*   Updated: 2025/05/05 15:40:22 by bgazur           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-int	ft_putstr(char *s)
+size_t	ft_putstr(char *s)
 {
 	size_t	length;
